Feat/admin

+++ server/controllers/images.js
@@ -9,11 +9,13 @@ const images = new Router({ prefix: '/images' })
 
 images.get('/', async ctx => {
   try {
-    const images = await Image.findAll()
+    const images = await Image.findAll({
+      attributes: ['id', 'path', 'alt']
+    })
     ctx.body = JSON.stringify(images)
   } catch (e) {
     ctx.status = 500
-    ctx.body = `Error ${e}`
+    ctx.body = JSON.stringify({ error: e })
   }
 })
 
@@ -23,7 +25,11 @@ images.post(
   async ctx => {
     try {
       const { files, fields } = await asyncBusboy(ctx.req)
-      const path = files.length ? await saveFile(files[0]) : null
+      const path =
+        files.length && files[0].filename.length
+          ? await saveFile(files[0])
+          : null
+      console.log(path)
       const image = await Image.create({
         path: path,
         alt: fields.alt || null,
@@ -33,7 +39,7 @@ images.post(
       ctx.body = JSON.stringify(image)
     } catch (e) {
       ctx.status = 500
-      ctx.body = `Error: ${e}`
+      ctx.body = JSON.stringify({ error: e })
     }
   }
 )
@@ -46,21 +52,36 @@ images.put(
       const id = ctx.params.id
       const { files, fields } = await asyncBusboy(ctx.req)
       const image = await Image.findById(id)
-      const path = files.length ? await saveFile(files[0]) : null
-      const savedImage = await Image.update(
+      const path =
+        files.length && files[0].filename.length
+          ? await saveFile(files[0])
+          : null
+      if (fields.cover_id) {
+        await Image.update(
+          { coverId: null },
+          { where: { albumId: image.albumId } }
+        )
+      }
+      await Image.update(
         {
           path: path || image.path,
           alt: fields.alt || image.alt,
-          albumId: fields.album_id || image.albumId
+          albumId: fields.album_id || image.albumId,
+          coverId: fields.cover_id || image.coverId
         },
-        { where: { id } },
-        { returning: true }
+        { where: { id } }
       )
-      await deleteFile(image.path)
+      const savedImage = await Image.findOne({
+        where: { id },
+        attributes: ['id', 'path', 'alt']
+      })
+      if (path) {
+        await deleteFile(image.path)
+      }
       ctx.body = JSON.stringify(savedImage)
     } catch (e) {
       ctx.status = 500
-      ctx.body = `Error: ${e}`
+      ctx.body = JSON.stringify({ error: e })
     }
   }
 )
@@ -78,7 +99,7 @@ images.delete(
       ctx.body = JSON.stringify(image)
     } catch (e) {
       ctx.status = 500
-      ctx.body = `Error: ${e}`
+      ctx.body = JSON.stringify({ error: e })
     }
   }
 )
